+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.js
@@ -7,6 +7,7 @@
 
 import React, { useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
+import { extractShapesArray } from '../../global/js/utils/props-helper';
 import { layout05, baseFontSize } from '@carbon/layout';
 import cx from 'classnames';
 import ReactResizeDetector from 'react-resize-detector';
@@ -28,12 +29,17 @@ import { ChevronUp16 } from '@carbon/icons-react';
 import {
   deprecateProp,
   deprecatePropUsage,
-  extractShapesArray,
   prepareProps,
 } from '../../global/js/utils/props-helper';
 
 const componentName = 'PageHeader';
-const blockClass = `${pkg.prefix}--page-header`;
+
+import {
+  blockClass,
+  utilCheckUpdateVerticalSpace,
+  utilGetTitleShape,
+  utilCalcSpacingBelowTitle,
+} from './PageHeaderUtils';
 
 export let PageHeader = React.forwardRef(
   (
@@ -60,60 +66,145 @@ export let PageHeader = React.forwardRef(
     },
     ref
   ) => {
-    const [hasActionBar, setHasActionBar] = useState(false);
-    const [actionBarItemArray, setActionBarItemArray] = useState([]);
-    const [pageActionsItemArray, setPageActionsItemArray] = useState([]);
     const [metrics, setMetrics] = useState({});
-    const [scrollYValue, setScrollYValue] = useState(0);
-    const [componentCssCustomProps, setComponentCssCustomProps] = useState({});
-    const [hasBreadcrumbRow, setHasBreadcrumbRow] = useState(false);
-    const [spacingBelowTitle, setSpacingBelowTitle] = useState('06');
-    const [
-      pageActionsInBreadcrumbRow,
-      setPageActionsInBreadcrumbRow,
-    ] = useState(false);
-    const [backgroundOpacity, setBackgroundOpacity] = useState(0);
-    const [hasCollapseButton, setHasCollapseButton] = useState(false);
-    const [spaceForCollapseButton, setSpaceForCollapseButton] = useState(false);
-    const [lastRowBufferActive, setLastRowBufferActive] = useState(false);
+
+    // refs
     const dynamicRefs = useRef({});
     const localHeaderRef = useRef(null);
     const headerRef = ref || localHeaderRef;
+
+    // utility functions
+    const calcSpacingBelowTitle = () =>
+      utilCalcSpacingBelowTitle(
+        availableSpace,
+        tags,
+        navigation,
+        subtitle,
+        pageActions
+      );
+    // Title shape is used to allow title to be string or shape
+    const getTitleShape = () =>
+      utilGetTitleShape(title, titleIcon, PageHeader.defaultProps.title);
+    const checkUpdateVerticalSpace = function () {
+      return utilCheckUpdateVerticalSpace(
+        headerRef,
+        navigation,
+        preventBreadcrumbScroll,
+        setMetrics,
+        dynamicRefs
+      );
+    };
+
+    // state based on props only
+    const actionBarItemArray = extractShapesArray(actionBarItems);
+    const hasActionBar = actionBarItemArray.length;
+    const hasBreadcrumbRow = !(
+      breadcrumbItems === undefined && actionBarItems === undefined
+    );
+    const pageActionsItemArray = extractShapesArray(pageActions)?.map(
+      (shape) => ({
+        label: shape.children,
+        ...shape,
+      })
+    );
+
+    const spacingBelowTitle = calcSpacingBelowTitle();
+
+    /* Title shape is used to allow title to be string or shape */
+    const titleShape = getTitleShape();
+
+    // NOTE: The buffer is used to add space between the bottom of the header and the last content
+    // No navigation and title row not pre-collapsed
+    // and zero or one of tags or (subtitle or available space)
+    const lastRowBufferActive =
+      !(navigation || tags) &&
+      (((title || pageActions) && !preCollapseTitleRow) ||
+        subtitle ||
+        availableSpace);
+
+    // state based on scroll/resize based effects
+    const [pageActionsInBreadcrumbRow, setPageActionsInBreadcrumbRow] =
+      useState(false);
+    const [scrollYValue, setScrollYValue] = useState(0);
+    const [componentCssCustomProps, setComponentCssCustomProps] = useState({});
+    const [backgroundOpacity, setBackgroundOpacity] = useState(0);
+    const [hasCollapseButton, setHasCollapseButton] = useState(false);
+    const [spaceForCollapseButton, setSpaceForCollapseButton] = useState(false);
     const [actionBarMaxWidth, setActionBarMaxWidth] = useState(0);
     const [actionBarMinWidth, setActionBarMinWidth] = useState(0);
-    const [
-      pageActionInBreadcrumbMaxWidth,
-      setPageActionInBreadcrumbMaxWidth,
-    ] = useState(0);
-    const [
-      pageActionInBreadcrumbMinWidth,
-      setPageActionInBreadcrumbMinWidth,
-    ] = useState(0);
+    const [pageActionInBreadcrumbMaxWidth, setPageActionInBreadcrumbMaxWidth] =
+      useState(0);
+    const [pageActionInBreadcrumbMinWidth, setPageActionInBreadcrumbMinWidth] =
+      useState(0);
     const [actionBarColumnWidth, setActionBarColumnWidth] = useState(0);
     const [fullyCollapsed, setFullyCollapsed] = useState(false);
 
-    /**
-     * * Title shape is used to allow title to be string or shape
-     */
-    const [titleShape, setTitleShape] = useState({});
-    useEffect(() => {
-      let newShape = { ...PageHeader.defaultProps.title };
+    // handlers
+    const handleActionBarWidthChange = ({ minWidth, maxWidth }) => {
+      /* don't know how to test resize */
+      /* istanbul ignore next */
+      setActionBarMaxWidth(maxWidth);
+      /* don't know how to test resize */
+      /* istanbul ignore next */
+      setActionBarMinWidth(minWidth);
+    };
 
-      if (title?.text) {
-        // title is in shape format
-        newShape = Object.assign(newShape, { ...title });
+    const handleButtonSetWidthChange = ({ minWidth, maxWidth }) => {
+      /* don't know how to test resize */
+      /* istanbul ignore next */
+      setPageActionInBreadcrumbMaxWidth(maxWidth);
+      /* don't know how to test resize */
+      /* istanbul ignore next */
+      setPageActionInBreadcrumbMinWidth(minWidth);
+    };
+
+    const handleResizeActionBarColumn = (width) => {
+      /* don't know how to test resize */
+      /* istanbul ignore next */
+      setActionBarColumnWidth(width);
+    };
+
+    const handleResize = () => {
+      // receives width and height parameters if needed
+      /* don't know how to test resize */
+      /* istanbul ignore next */
+      checkUpdateVerticalSpace();
+    };
+
+    const toggleCollapse = (forceCollapse) => {
+      const collapse =
+        typeof forceCollapse !== 'undefined' ? forceCollapse : !fullyCollapsed;
+
+      /* don't know how to test resize */
+      /* istanbul ignore next if */
+      if (collapse) {
+        window.scrollTo({
+          top: pageHeaderOffset - (metrics?.headerTopValue || 0),
+          behavior: 'smooth',
+        });
       } else {
-        // title is a string
-        newShape.text = title;
+        window.scrollTo({ top: 0, behavior: 'smooth' });
       }
+    };
 
-      if (!newShape.icon && titleIcon) {
-        // if no icon use titleIcon if supplied
-        newShape.icon = titleIcon;
-      }
+    const handleCollapseToggle = () => {
+      toggleCollapse();
+    };
 
-      setTitleShape(newShape);
-    }, [title, titleIcon]);
+    // use effects
+    useEffect(() => {
+      // Determine the location of the pageAction buttons
+      setPageActionsInBreadcrumbRow(
+        preCollapseTitleRow ||
+          (scrollYValue > metrics.titleRowSpaceAbove && hasActionBar)
+      );
+    }, [
+      hasActionBar,
+      metrics.breadcrumbRowSpaceBelow,
+      metrics.titleRowSpaceAbove,
+      preCollapseTitleRow,
+      scrollYValue,
+    ]);
 
     useEffect(() => {
       let newActionBarWidth = 'initial';
@@ -160,149 +251,6 @@ export let PageHeader = React.forwardRef(
       actionBarMinWidth,
       pageActionInBreadcrumbMaxWidth,
       pageActionInBreadcrumbMinWidth,
-    ]);
-
-    const handleActionBarWidthChange = ({ minWidth, maxWidth }) => {
-      /* istanbul ignore next */ /* don't know how to test resize */
-      setActionBarMaxWidth(maxWidth); /* don't know how to test resize */
-      /* istanbul ignore next */ setActionBarMinWidth(minWidth);
-    };
-
-    const handleButtonSetWidthChange = ({ minWidth, maxWidth }) => {
-      setPageActionInBreadcrumbMaxWidth(maxWidth);
-      setPageActionInBreadcrumbMinWidth(minWidth);
-    };
-
-    const handleResizeActionBarColumn = (width) => {
-      /* istanbul ignore next */ /* don't know how to test resize */
-      setActionBarColumnWidth(width);
-    };
-
-    const getDynamicRef = (selector) => {
-      // would love to do this differently but digging in the dom seems easier
-      // than getting a ref to a conditionally rendered item
-      /* istanbul ignore next if */ /* don't know how to test resize */
-      if (!headerRef.current) {
-        return undefined;
-      } else {
-        let dRef = dynamicRefs.current[selector];
-        if (!dRef || dRef.parentNode === null) {
-          dynamicRefs.current[selector] = headerRef.current.querySelector(
-            selector
-          );
-        }
-      }
-      return dynamicRefs.current[selector];
-    };
-
-    const checkUpdateVerticalSpace = () => {
-      // Utility function that checks the heights of various elements which are used to determine layout
-      const update = {};
-
-      const breadcrumbTitleEl = getDynamicRef(
-        `.${blockClass}__breadcrumb-title`
-      );
-      const breadcrumbRowEl = getDynamicRef(`.${blockClass}__breadcrumb-row`);
-      const titleRowEl = getDynamicRef(`.${blockClass}__title-row`);
-      const subtitleRowEl = getDynamicRef(`.${blockClass}__subtitle-row`);
-      const availableRowEl = getDynamicRef(`.${blockClass}__available-row`);
-      const navigationRowEl = getDynamicRef(`.${blockClass}__navigation-row`);
-
-      update.headerHeight = headerRef.current
-        ? headerRef.current.clientHeight
-        : 0;
-      update.headerWidth = headerRef.current
-        ? headerRef.current.offsetWidth
-        : 0;
-
-      update.breadcrumbRowHeight = breadcrumbRowEl
-        ? breadcrumbRowEl.clientHeight
-        : 0;
-      update.breadcrumbRowWidth = breadcrumbRowEl
-        ? breadcrumbRowEl.offsetWidth
-        : 0;
-
-      update.breadcrumbTitleHeight = breadcrumbTitleEl
-        ? breadcrumbTitleEl.clientHeight
-        : 1;
-
-      update.titleRowHeight = titleRowEl ? titleRowEl.clientHeight : 0;
-      update.subtitleRowHeight = subtitleRowEl ? subtitleRowEl.clientHeight : 0;
-      update.availableRowHeight = availableRowEl
-        ? availableRowEl.clientHeight
-        : 0;
-      update.navigationRowHeight = navigationRowEl
-        ? navigationRowEl.clientHeight
-        : 1;
-
-      update.breadcrumbRowSpaceBelow = 0;
-      update.titleRowSpaceAbove = 0;
-
-      update.headerTopValue = navigation
-        ? preventBreadcrumbScroll
-          ? update.navigationRowHeight +
-            update.breadcrumbRowHeight -
-            update.headerHeight
-          : update.navigationRowHeight - update.headerHeight
-        : update.breadcrumbRowHeight - update.headerHeight;
-
-      if (window) {
-        let val;
-        if (breadcrumbRowEl) {
-          val = parseFloat(
-            window
-              .getComputedStyle(breadcrumbRowEl)
-              .getPropertyValue('margin-bottom'),
-            10
-          );
-          update.breadcrumbRowSpaceBelow = isNaN(val) ? 0 : val;
-        }
-        if (titleRowEl) {
-          val = parseFloat(
-            window.getComputedStyle(titleRowEl).getPropertyValue('margin-top'),
-            10
-          );
-          update.titleRowSpaceAbove = isNaN(val) ? 0 : val;
-        }
-      }
-
-      setMetrics((previous) => ({ ...previous, ...update }));
-    };
-
-    useEffect(() => {
-      // NOTE: The buffer is used to add space between the bottom of the header and the last content
-
-      // No navigation and title row not pre-collapsed
-      // and zero or one of tags or (subtitle or available space)
-      setLastRowBufferActive(
-        !(navigation || tags) &&
-          (((title || pageActions) && !preCollapseTitleRow) ||
-            subtitle ||
-            availableSpace)
-      );
-    }, [
-      availableSpace,
-      navigation,
-      pageActions,
-      preCollapseTitleRow,
-      setLastRowBufferActive,
-      subtitle,
-      tags,
-      title,
-    ]);
-
-    useEffect(() => {
-      // Determine the location of the pageAction buttons
-      setPageActionsInBreadcrumbRow(
-        preCollapseTitleRow ||
-          (scrollYValue > metrics.titleRowSpaceAbove && hasActionBar)
-      );
-    }, [
-      hasActionBar,
-      metrics.breadcrumbRowSpaceBelow,
-      metrics.titleRowSpaceAbove,
-      preCollapseTitleRow,
-      scrollYValue,
     ]);
 
     useEffect(() => {
@@ -418,47 +366,6 @@ export let PageHeader = React.forwardRef(
     ]);
 
     useEffect(() => {
-      // Breadcrumb row only rendered if true
-      // eslint-disable-next-line
-      setHasBreadcrumbRow(
-        !(breadcrumbItems === undefined && actionBarItems === undefined)
-      );
-    }, [actionBarItems, breadcrumbItems]);
-
-    useEffect(() => {
-      const newShapes = extractShapesArray(actionBarItems);
-      setHasActionBar(newShapes.length);
-      setActionBarItemArray(newShapes);
-    }, [actionBarItems]);
-
-    useEffect(() => {
-      const shapes = extractShapesArray(pageActions);
-      setPageActionsItemArray(
-        shapes?.map((shape) => ({ label: shape.children, ...shape }))
-      );
-    }, [pageActions]);
-
-    useEffect(() => {
-      // Determines the amount of space needed below the title
-      let belowTitleSpace = 'default';
-
-      if (
-        pageActions !== undefined &&
-        navigation !== undefined &&
-        subtitle === undefined &&
-        availableSpace === undefined
-      ) {
-        belowTitleSpace = '06';
-      } else if (subtitle !== undefined || availableSpace !== undefined) {
-        belowTitleSpace = '03';
-      } else if (navigation === undefined && tags !== undefined) {
-        belowTitleSpace = '05';
-      }
-
-      setSpacingBelowTitle(belowTitleSpace);
-    }, [availableSpace, tags, navigation, subtitle, pageActions]);
-
-    useEffect(() => {
       // Determines if the background should be one based on the header height or scroll
       let result = background && 1;
 
@@ -469,6 +376,9 @@ export let PageHeader = React.forwardRef(
       ) {
         const startAddingAt = parseFloat(layout05, 10) * parseInt(baseFontSize);
         const scrollRemaining = metrics.headerHeight - scrollYValue;
+
+        /* don't know how to test resize */
+        /* istanbul ignore next if */
         if (scrollRemaining < startAddingAt) {
           const distanceAddingOver =
             startAddingAt - metrics.breadcrumbRowHeight;
@@ -498,41 +408,17 @@ export let PageHeader = React.forwardRef(
     ]);
 
     useEffect(() => {
+      // Determine if space is needed in the breadcrumb for a collapse button
       setSpaceForCollapseButton(
         hasCollapseButton && !(navigation || tags) && metrics.headerHeight
       );
     }, [hasCollapseButton, navigation, tags, metrics.headerHeight]);
-
-    const handleResize = () => {
-      // receives width and height parameters if needed
-      /* don't know how to test resize */
-      /* istanbul ignore next */
-      checkUpdateVerticalSpace();
-    };
 
     const nextToTabsCheck = () => {
       return (
         actionBarItems === undefined &&
         scrollYValue + metrics.headerTopValue > 0
       );
-    };
-
-    const toggleCollapse = (forceCollapse) => {
-      const collapse =
-        typeof forceCollapse !== 'undefined' ? forceCollapse : !fullyCollapsed;
-
-      if (collapse) {
-        window.scrollTo({
-          top: pageHeaderOffset - (metrics?.headerTopValue || 0),
-          behavior: 'smooth',
-        });
-      } else {
-        window.scrollTo({ top: 0, behavior: 'smooth' });
-      }
-    };
-
-    const handleCollapseToggle = () => {
-      toggleCollapse();
     };
 
     useEffect(() => {
@@ -564,9 +450,12 @@ export let PageHeader = React.forwardRef(
             {hasBreadcrumbRow ? (
               <Row
                 className={cx(`${blockClass}__breadcrumb-row`, {
-                  [`${blockClass}__breadcrumb-row--next-to-tabs`]: nextToTabsCheck(),
-                  [`${blockClass}__breadcrumb-row--has-breadcrumbs`]: breadcrumbItems,
-                  [`${blockClass}__breadcrumb-row--has-action-bar`]: hasActionBar,
+                  [`${blockClass}__breadcrumb-row--next-to-tabs`]:
+                    nextToTabsCheck(),
+                  [`${blockClass}__breadcrumb-row--has-breadcrumbs`]:
+                    breadcrumbItems,
+                  [`${blockClass}__breadcrumb-row--has-action-bar`]:
+                    hasActionBar,
                 })}>
                 <div className={`${blockClass}__breadcrumb-row--container`}>
                   <Column
@@ -588,7 +477,8 @@ export let PageHeader = React.forwardRef(
                             className={cx([
                               `${blockClass}__breadcrumb-title`,
                               {
-                                [`${blockClass}__breadcrumb-title--pre-collapsed`]: preCollapseTitleRow,
+                                [`${blockClass}__breadcrumb-title--pre-collapsed`]:
+                                  preCollapseTitleRow,
                               },
                             ])}>
                             {titleLoading ? <SkeletonText /> : titleText}
@@ -605,8 +495,10 @@ export let PageHeader = React.forwardRef(
                     className={cx([
                       `${blockClass}__action-bar-column ${blockClass}__action-bar-column--background`,
                       {
-                        [`${blockClass}__action-bar-column--has-page-actions`]: pageActions,
-                        [`${blockClass}__action-bar-column--influenced-by-collapse-button`]: spaceForCollapseButton,
+                        [`${blockClass}__action-bar-column--has-page-actions`]:
+                          pageActions,
+                        [`${blockClass}__action-bar-column--influenced-by-collapse-button`]:
+                          spaceForCollapseButton,
                       },
                     ])}>
                     <ReactResizeDetector
@@ -620,7 +512,8 @@ export let PageHeader = React.forwardRef(
                             {pageActions && (
                               <div
                                 className={cx(`${blockClass}__page-actions`, {
-                                  [`${blockClass}__page-actions--in-breadcrumb`]: pageActionsInBreadcrumbRow,
+                                  [`${blockClass}__page-actions--in-breadcrumb`]:
+                                    pageActionsInBreadcrumbRow,
                                 })}>
                                 <ButtonSetWithOverflow
                                   className={`${blockClass}__button-set--in-breadcrumb`}
@@ -651,8 +544,10 @@ export let PageHeader = React.forwardRef(
                   `${blockClass}__title-row`,
                   `${blockClass}__title-row--spacing-below-${spacingBelowTitle}`,
                   {
-                    [`${blockClass}__title-row--no-breadcrumb-row`]: !hasBreadcrumbRow,
-                    [`${blockClass}__title-row--under-action-bar`]: hasActionBar,
+                    [`${blockClass}__title-row--no-breadcrumb-row`]:
+                      !hasBreadcrumbRow,
+                    [`${blockClass}__title-row--under-action-bar`]:
+                      hasActionBar,
                     [`${blockClass}__title-row--sticky`]:
                       pageActions !== undefined &&
                       actionBarItems === undefined &&
@@ -685,7 +580,8 @@ export let PageHeader = React.forwardRef(
                 {pageActions !== undefined ? (
                   <Column
                     className={cx(`${blockClass}__page-actions`, {
-                      [`${blockClass}__page-actions--in-breadcrumb`]: pageActionsInBreadcrumbRow,
+                      [`${blockClass}__page-actions--in-breadcrumb`]:
+                        pageActionsInBreadcrumbRow,
                     })}>
                     <ButtonSetWithOverflow
                       className={`${blockClass}__page-actions-container`}
@@ -726,7 +622,8 @@ export let PageHeader = React.forwardRef(
                 className={cx([
                   `${blockClass}__last-row-buffer`,
                   {
-                    [`${blockClass}__last-row-buffer--active`]: lastRowBufferActive,
+                    [`${blockClass}__last-row-buffer--active`]:
+                      lastRowBufferActive,
                   },
                 ])}></div>
             )}
@@ -762,7 +659,8 @@ export let PageHeader = React.forwardRef(
           {hasCollapseButton ? (
             <Button
               className={cx(`${blockClass}__collapse-expand-toggle`, {
-                [`${blockClass}__collapse-expand-toggle--collapsed`]: fullyCollapsed,
+                [`${blockClass}__collapse-expand-toggle--collapsed`]:
+                  fullyCollapsed,
               })}
               data-collapse={fullyCollapsed ? 'collapsed' : 'not collapsed'}
               hasIconOnly={true}