+++ packages/jest-config/setup/setupFilesAfterEnv.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2020, 2020
+ * Copyright IBM Corp. 2020, 2021
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,12 +7,10 @@
 
 // `setupFilesAfterEnv` enables running the code immediately after the test framework has been installed in the environment - https://jestjs.io/docs/en/configuration.html#setupfilesafterenv-array
 
-// **TODO** restore 12-13 & 32-73 and remove 75-81 once carbon fixes issues that fail our tests
-
 import '@testing-library/jest-dom';
 
-// import chalk from 'chalk';
-// import util from 'util';
+import chalk from 'chalk';
+import util from 'util';
 
 import toHaveNoAxeViolations from './matchers/toHaveNoAxeViolations';
 import toBeAccessible from './matchers/toBeAccessible';
@@ -31,51 +29,48 @@ expect.extend({ toBeAccessible, toHaveNoAxeViolations });
 //
 // Inspired by the following setup from facebook/react
 // https://github.com/facebook/react/blob/6250462bed19c9f18a8cf3c2b5fcaf9aba1df72b/scripts/jest/setupTests.js#L69
-// const consoleMethods = ['error', 'warn', process.env.CI && 'log'].filter(
-//   Boolean
-// );
 
-// for (const methodName of consoleMethods) {
-//   const unexpectedConsoleCallStacks = [];
-//   const newMethod = function (format, ...args) {
-//     const stack = new Error().stack;
-//     unexpectedConsoleCallStacks.push([
-//       stack.substr(stack.indexOf('\n') + 1),
-//       util.format(format, ...args),
-//     ]);
-//   };
+const oldConsole = {};
+['error', 'warn', process.env.CI && 'log'].filter(Boolean).forEach((mthd) => {
+  const unexpectedConsoleCallStacks = [];
 
-//   console[methodName] = newMethod;
+  oldConsole[mthd] = console[mthd];
+  const newMethod = function (format, ...args) {
+    const stack = new Error().stack;
+    unexpectedConsoleCallStacks.push([
+      stack.substr(stack.indexOf('\n') + 1),
+      util.format(format, ...args),
+    ]);
+  };
+  console[mthd] = newMethod;
 
-//   global.beforeEach(() => {
-//     unexpectedConsoleCallStacks.length = 0;
-//   });
+  global.beforeEach(() => {
+    unexpectedConsoleCallStacks.length = 0;
+  });
 
-//   global.afterEach(() => {
-//     if (console[methodName] !== newMethod) {
-//       throw new Error(`Test did not restore a mock for console.${methodName}`);
-//     }
+  global.afterEach(() => {
+    if (console[mthd] !== newMethod) {
+      //throw new Error(
+      //  `Test did not restore a mock for ${chalk.bold(`console.${mthd}()`)}`
+      //);
+    }
 
-//     if (unexpectedConsoleCallStacks.length > 0) {
-//       const messages = unexpectedConsoleCallStacks.map(
-//         ([stack, message]) =>
-//           `${message}\n` +
-//           `${stack
-//             .split('\n')
-//             .map((line) => chalk.gray(line))
-//             .join('\n')}`
-//       );
-//       const message = `Expected test not to call ${chalk.bold(
-//         `console.${methodName}()`
-//       )}`;
+    if (unexpectedConsoleCallStacks.length > 0) {
+      const messages = unexpectedConsoleCallStacks.map(
+        ([stack, message]) =>
+          `${message}\n` +
+          `${stack
+            .split('\n')
+            .map((line) => chalk.gray(line))
+            .join('\n')}`
+      );
+      const message = `Expected test not to call ${chalk.bold(
+        `console.${mthd}()`
+      )}`;
 
-//       throw new Error(`${message}\n\n${messages.join('\n\n')}`);
-//     }
-//   });
-// }
-
-global.beforeEach(() => {
-  jest.spyOn(global.console, 'warn');
+      throw new Error(`${message}\n\n${messages.join('\n\n')}`);
+    }
+  });
 });
 
 global.afterEach(() => {